Change: authentication when nodes rejoin network

+++ sls_cli.h
@@ -68,7 +68,7 @@ struct node_db_struct_t {
 	uint8_t 	last_emergency_msg[MAX_CMD_DATA_LEN];
 	char		next_hop_addr[16];			//byte array
 	char		next_hop_link_addr[40];		// link address string
-	double		delay;
+	float		delay;
 	double		rdr;	//request delivery rate
 	uint16_t	challenge_code;
 	uint16_t	challenge_code_res;
